# Use forward declarations instead of IWYU