--- a/Firestore/Source/API/FIRCollectionReference+Internal.h
+++ b/Firestore/Source/API/FIRCollectionReference+Internal.h
@@ -18,8 +18,15 @@
 
 #include <memory>
 
-#include "Firestore/core/src/firebase/firestore/api/collection_reference.h"
-#include "Firestore/core/src/firebase/firestore/model/resource_path.h"
+#include "Firestore/core/src/firebase/firestore/api/api_fwd.h"
+
+namespace firebase {
+namespace firestore {
+namespace model {
+class ResourcePath;
+}  // namespace model
+}  // namespace firestore
+}  // namespace firebase
 
 namespace api = firebase::firestore::api;
 namespace model = firebase::firestore::model;