--- a/Firestore/Source/API/FIRDocumentReference+Internal.h
+++ b/Firestore/Source/API/FIRDocumentReference+Internal.h
@@ -18,9 +18,15 @@
 
 #include <memory>
 
-#include "Firestore/core/src/firebase/firestore/api/document_reference.h"
-#include "Firestore/core/src/firebase/firestore/model/document_key.h"
-#include "Firestore/core/src/firebase/firestore/model/resource_path.h"
+#include "Firestore/core/src/firebase/firestore/api/api_fwd.h"
+namespace firebase {
+namespace firestore {
+namespace model {
+class DocumentKey;
+class ResourcePath;
+}  // namespace model
+}  // namespace firestore
+}  // namespace firebase
 
 namespace api = firebase::firestore::api;
 namespace model = firebase::firestore::model;