--- a/Firestore/Source/API/FIRFieldPath+Internal.h
+++ b/Firestore/Source/API/FIRFieldPath+Internal.h
@@ -16,7 +16,7 @@
 
 #import "FIRFieldPath.h"
 
-#include "Firestore/core/src/firebase/firestore/model/field_path.h"
+#include "Firestore/core/src/firebase/firestore/model/model_fwd.h"
 
 namespace model = firebase::firestore::model;