--- a/Firestore/Source/API/FIRQuerySnapshot+Internal.h
+++ b/Firestore/Source/API/FIRQuerySnapshot+Internal.h
@@ -18,10 +18,8 @@
 
 #include <memory>
 
-#include "Firestore/core/src/firebase/firestore/api/firestore.h"
-#include "Firestore/core/src/firebase/firestore/api/query_snapshot.h"
-#include "Firestore/core/src/firebase/firestore/api/snapshot_metadata.h"
-#include "Firestore/core/src/firebase/firestore/core/view_snapshot.h"
+#include "Firestore/core/src/firebase/firestore/api/api_fwd.h"
+#include "Firestore/core/src/firebase/firestore/core/core_fwd.h"
 
 @class FIRFirestore;
 @class FIRSnapshotMetadata;