--- a/Firestore/core/src/firebase/firestore/api/collection_reference.h
+++ b/Firestore/core/src/firebase/firestore/api/collection_reference.h
@@ -20,21 +20,19 @@
 #include <memory>
 #include <string>
 
+#include "Firestore/core/src/firebase/firestore/api/api_fwd.h"
 #include "Firestore/core/src/firebase/firestore/api/query_core.h"
-#include "absl/strings/string_view.h"
+#include "Firestore/core/src/firebase/firestore/core/core_fwd.h"
 #include "absl/types/optional.h"
 
 namespace firebase {
 namespace firestore {
-namespace core {
 
-class ParsedSetData;
-
-}  // namespace core
+namespace model {
+class ResourcePath;
+}  // namespace model
 
 namespace api {
-
-class DocumentReference;
 
 /**
  * A `CollectionReference` object can be used for adding documents, getting