--- a/Firestore/core/src/firebase/firestore/api/document_reference.cc
+++ b/Firestore/core/src/firebase/firestore/api/document_reference.cc
@@ -110,7 +110,7 @@ void DocumentReference::DeleteDocument(util::StatusCallback callback) {
 }
 
 void DocumentReference::GetDocument(Source source,
-                                    DocumentSnapshot::Listener&& callback) {
+                                    DocumentSnapshotListener&& callback) {
   if (source == Source::Cache) {
     firestore_->client()->GetDocumentFromLocalCache(*this, std::move(callback));
     return;
@@ -123,7 +123,7 @@ void DocumentReference::GetDocument(Source source,
 
   class ListenOnce : public EventListener<DocumentSnapshot> {
    public:
-    ListenOnce(Source source, DocumentSnapshot::Listener&& listener)
+    ListenOnce(Source source, DocumentSnapshotListener&& listener)
         : source_(source), listener_(std::move(listener)) {
     }
 
@@ -173,7 +173,7 @@ void DocumentReference::GetDocument(Source source,
 
    private:
     Source source_;
-    DocumentSnapshot::Listener listener_;
+    DocumentSnapshotListener listener_;
 
     std::promise<std::unique_ptr<ListenerRegistration>> registration_promise_;
   };
@@ -187,12 +187,12 @@ void DocumentReference::GetDocument(Source source,
 }
 
 std::unique_ptr<ListenerRegistration> DocumentReference::AddSnapshotListener(
-    ListenOptions options, DocumentSnapshot::Listener&& user_listener) {
+    ListenOptions options, DocumentSnapshotListener&& user_listener) {
   // Convert from ViewSnapshots to DocumentSnapshots.
   class Converter : public EventListener<ViewSnapshot> {
    public:
     Converter(DocumentReference* parent,
-              DocumentSnapshot::Listener&& user_listener)
+              DocumentSnapshotListener&& user_listener)
         : firestore_(parent->firestore_),
           key_(parent->key_),
           user_listener_(std::move(user_listener)) {
@@ -224,7 +224,7 @@ std::unique_ptr<ListenerRegistration> DocumentReference::AddSnapshotListener(
    private:
     std::shared_ptr<Firestore> firestore_;
     DocumentKey key_;
-    DocumentSnapshot::Listener user_listener_;
+    DocumentSnapshotListener user_listener_;
   };
   auto view_listener =
       absl::make_unique<Converter>(this, std::move(user_listener));