--- a/Firestore/core/src/firebase/firestore/api/document_reference.h
+++ b/Firestore/core/src/firebase/firestore/api/document_reference.h
@@ -21,29 +21,19 @@
 #include <string>
 #include <utility>
 
-#include "Firestore/core/src/firebase/firestore/api/document_snapshot.h"
-#include "Firestore/core/src/firebase/firestore/api/listener_registration.h"
-#include "Firestore/core/src/firebase/firestore/core/listen_options.h"
+#include "Firestore/core/src/firebase/firestore/api/api_fwd.h"
+#include "Firestore/core/src/firebase/firestore/core/core_fwd.h"
 #include "Firestore/core/src/firebase/firestore/model/document_key.h"
-#include "Firestore/core/src/firebase/firestore/model/resource_path.h"
-#include "Firestore/core/src/firebase/firestore/util/nullability.h"
-#include "Firestore/core/src/firebase/firestore/util/status.h"
 #include "Firestore/core/src/firebase/firestore/util/status_fwd.h"
 
 namespace firebase {
 namespace firestore {
-namespace core {
 
-class ParsedSetData;
-class ParsedUpdateData;
-
-}  // namespace core
+namespace model {
+class ResourcePath;
+}  // namespace model
 
 namespace api {
-
-class CollectionReference;
-class Firestore;
-enum class Source;
 
 class DocumentReference {
  public: