--- a/Firestore/core/src/firebase/firestore/api/document_snapshot.cc
+++ b/Firestore/core/src/firebase/firestore/api/document_snapshot.cc
@@ -18,6 +18,7 @@
 
 #include "Firestore/core/src/firebase/firestore/api/document_reference.h"
 #include "Firestore/core/src/firebase/firestore/model/field_value.h"
+#include "Firestore/core/src/firebase/firestore/model/resource_path.h"
 #include "Firestore/core/src/firebase/firestore/util/hashing.h"
 #include "absl/types/optional.h"