--- a/Firestore/core/src/firebase/firestore/api/document_snapshot.h
+++ b/Firestore/core/src/firebase/firestore/api/document_snapshot.h
@@ -38,8 +38,6 @@ class Firestore;
 
 class DocumentSnapshot {
  public:
-  using Listener = std::unique_ptr<core::EventListener<DocumentSnapshot>>;
-
   DocumentSnapshot() = default;
 
   static DocumentSnapshot FromDocument(std::shared_ptr<Firestore> firestore,
@@ -89,6 +87,9 @@ class DocumentSnapshot {
   absl::optional<model::Document> internal_document_;
   SnapshotMetadata metadata_;
 };
+
+using DocumentSnapshotListener =
+    std::unique_ptr<core::EventListener<DocumentSnapshot>>;
 
 inline bool operator!=(const DocumentSnapshot& lhs,
                        const DocumentSnapshot& rhs) {