--- a/Firestore/core/src/firebase/firestore/api/listener_registration.h
+++ b/Firestore/core/src/firebase/firestore/api/listener_registration.h
@@ -19,9 +19,10 @@
 
 namespace firebase {
 namespace firestore {
+
 namespace core {
 class FirestoreClient;
-}
+}  // namespace core
 
 namespace api {