--- a/Firestore/core/src/firebase/firestore/api/query_core.cc
+++ b/Firestore/core/src/firebase/firestore/api/query_core.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google
+ * Copyright 2019 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.