… .github/workflows/core.yml
… .github/workflows/dynamiclinks.yml
… .github/workflows/firestore.yml
… .github/workflows/storage.yml
… Firestore/Example/Tests/API/FIRFieldPathTests.mm
… Firestore/Example/Tests/API/FIRQueryUnitTests.mm
… Firestore/Example/Tests/Integration/API/FIRDatabaseTests.mm
… Firestore/Example/Tests/Integration/FSTDatastoreTests.mm
… Firestore/Example/Tests/SpecTests/FSTMockDatastore.h
… Firestore/Example/Tests/SpecTests/FSTSpecTests.mm
… Firestore/Example/Tests/SpecTests/FSTSyncEngineTestDriver.h
… Firestore/Example/Tests/SpecTests/FSTSyncEngineTestDriver.mm
… Firestore/Example/Tests/Util/FSTHelpers.h
… Firestore/Example/Tests/Util/FSTHelpers.mm
… Firestore/core/src/firebase/firestore/api/CMakeLists.txt
… Firestore/core/src/firebase/firestore/api/query_listener_registration.h
… Firestore/core/src/firebase/firestore/api/query_snapshot.h
… Firestore/core/src/firebase/firestore/api/snapshots_in_sync_listener_registration.cc
… Firestore/core/src/firebase/firestore/api/snapshots_in_sync_listener_registration.h
… Firestore/core/src/firebase/firestore/api/write_batch.cc
… Firestore/core/src/firebase/firestore/api/write_batch.h
… Firestore/core/src/firebase/firestore/auth/credentials_provider.cc
… Firestore/core/src/firebase/firestore/auth/token.cc
… Firestore/core/src/firebase/firestore/auth/token.h
… Firestore/core/src/firebase/firestore/core/CMakeLists.txt
… Firestore/core/src/firebase/firestore/core/array_contains_any_filter.h
… Firestore/core/src/firebase/firestore/core/array_contains_filter.cc
… Firestore/core/src/firebase/firestore/core/array_contains_filter.h
… Firestore/core/src/firebase/firestore/core/bound.cc
… Firestore/core/src/firebase/firestore/core/bound.h