On my MBP, this results in a 15% improvement in build time for Firestore proper (excluding tests), from ~36 sec to ~30 sec. This is a user-visible improvement.

More importantly though, the blast radius of many changes is greatly reduced. For example, making a one-line change to these headers results in a 35-45% reduction in incremental rebuild time:

Filename              | Objects            | Time         | Improvement
--------------------- | ------------------ | ------------ | -----------
model/document.h      | 150 ->  81 objects | 47 -> 25 sec | 45% faster
model/field_value.h   | 159 -> 146 objects | 80 -> 43 sec | 45% faster
model/resource_path.h | 176 -> 102 objects | 52 -> 34 sec | 35% faster
core/query.h          | 106 ->  63 objects | 38 -> 25 sec | 35% faster

It's probably best to review this a commit at a time. Interesting changes have been factored out separately. The last commit is mostly the mechanical changes.